frameDuratino—>UInt16

+++ Griffy/BluetoothManager.swift
@@ -98,7 +98,7 @@ final class BluetoothManager: NSObject {
     if let animation = GFCharacteristic.animation, let frameCount = GFCharacteristic.frameCount, let frameDuration = GFCharacteristic.frameDuration {
       writeValue(data: UInt8(griffy.radialFilePaths.count == 1 ? 0 : 1).data, toCharacteristic: animation)
       writeValue(data: UInt8(griffy.radialFilePaths.count).data, toCharacteristic: frameCount)
-      writeValue(data: UInt8(griffy.frameDuration).data, toCharacteristic: frameDuration)
+      writeValue(data: UInt16(griffy.frameDuration).data, toCharacteristic: frameDuration)
     }
     
     for radial in griffy.radialFilePaths {
